+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - feature: Charger Mode is supported (electrical.chargers.<#>.mode)
 - feature: Charger State of Charge is supported (electrical.chargers.<#>.capacity.stateOfCharge)
 
+- feature: EmpirBus NXT switches defined as special device type `leakSensor` are shown as a leak sensor
+
 - fix: Cleanup HomeKit device meta data for new devices. Should not have side effects on existing devices.
 
 **Known issues:**